depreciation

+++ Vehicle.java
@@ -2,10 +2,10 @@ public abstract class Vehicle {
     protected String make;
     protected String model;
     protected int year;
-    protected double price;
+    protected int price;
     protected int quantity;
 
-    public Vehicle(String make, String model, int year, double price, int quantity) {
+    public Vehicle(String make, String model, int year, int price, int quantity) {
         this.make = make;
         this.model = model;
         this.year = year;
@@ -13,7 +13,7 @@ public abstract class Vehicle {
         this.quantity = quantity;
     }
 
-    public abstract void displayVehicleDetails(); 
+    public abstract void displayVehicleDetails();
 
     public String getMake() {
         return this.make;
@@ -39,11 +39,11 @@ public abstract class Vehicle {
         this.year = year;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return this.price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
